Set up abstract controllers for the API usage.
Most of this is actually from the modules but I need
my controllers to extend the base hyla controllers so
I copied most of them. I will be changing the API
controller code quite a bit simply because I hate the
'redirect' in before() hack and want my routes to
handle that.